+++ src/OpenTelemetry.Instrumentation.ConfluentKafka/ConfluentKafkaCommon.cs
@@ -17,8 +17,8 @@ namespace OpenTelemetry.Instrumentation.ConfluentKafka;
 /// Contains common constants and static members used by the Confluent Kafka instrumentation.
 /// </summary>
 /// <remarks>
-/// Follows the v1.43.0 messaging semantic conventions:
-/// https://github.com/open-telemetry/semantic-conventions/tree/v1.43.0/docs/messaging.
+/// Follows the v1.44.0 messaging semantic conventions:
+/// https://github.com/open-telemetry/semantic-conventions/tree/v1.44.0/docs/messaging.
 /// </remarks>
 internal static class ConfluentKafkaCommon
 {
@@ -34,7 +34,7 @@ internal static class ConfluentKafkaCommon
     internal const string ReceiveOperationType = "receive";
     internal const string ProcessOperationType = "process";
 
-    internal static readonly Version SemanticConventionsVersion = new(1, 43, 0);
+    internal static readonly Version SemanticConventionsVersion = new(1, 44, 0);
 
     internal static readonly ActivitySource ActivitySource = ActivitySourceFactory.Create(typeof(ConfluentKafkaCommon), SemanticConventionsVersion);
     internal static readonly Meter Meter = MeterFactory.Create(typeof(ConfluentKafkaCommon), SemanticConventionsVersion);