+++ src/OpenTelemetry.Instrumentation.ConfluentKafka/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Updated Semantic Conventions to [v1.44.0](https://github.com/open-telemetry/semantic-conventions/releases/tag/v1.44.0).
+  ([#4739](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4739))
+
 * Added `messaging.kafka.cluster.id` span attribute.
   ([#4739](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4739))