+++ src/OpenTelemetry.Instrumentation.ConfluentKafka/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Updated Semantic Conventions to [v1.44.0](https://github.com/open-telemetry/semantic-conventions/releases/tag/v1.44.0).
+  ([#4976](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4976))
+
 ## 0.2.0-alpha.2
 
 Released 2026-Jul-17