+++ src/OpenTelemetry.Instrumentation.ConfluentKafka/CHANGELOG.md
@@ -8,6 +8,9 @@
 * Added `messaging.kafka.cluster.id` span attribute.
   ([#4739](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4739))
 
+* Updated OpenTelemetry core component version(s) to `1.18.0`.
+  ([#5022](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/5022))
+
 ## 0.2.0-alpha.2
 
 Released 2026-Jul-17